requirements.txt
pandas==2.2.3
matplotlib==3.9.2
numpy==2.1.3
reportlab==4.2.5
pillow==11.0.0